Test for descendant element delete permissions before deleting an element container

+++ src/Umbraco.Cms.Api.Management/Controllers/Element/Folder/DeleteElementFolderController.cs
@@ -57,6 +57,17 @@ public class DeleteElementFolderController : ElementFolderControllerBase
             return Forbidden();
         }
 
+        // Also authorize deletion of all descendant elements.
+        authorizationResult = await _authorizationService.AuthorizeResourceAsync(
+            User,
+            ElementPermissionResource.Branch(ActionElementDelete.ActionLetter, id),
+            AuthorizationPolicies.ElementPermissionByResource);
+
+        if (!authorizationResult.Succeeded)
+        {
+            return Forbidden();
+        }
+
         return await DeleteFolderAsync(id);
     }
 }
